:QLite: EXPERIMENTAL: turn on "SQLITE_NO_SYNC"
It's been observed at one of my clients that a huge source of
short writes to the EMMC (which are responsible for lag, stutter,
and can even lead to wearing-out of the Flash device) are caused
by SQLite DB meta-writes that are (f)sync()ed out. We have done
some testing on this end to examine what damage that might occur
if we don't sync all DB metadata, and have decided that the
incidence of sudden battery loss or kernel panic()s is small.

Note: the change to fdatasync() for fsync() in JB helps, but this
problem still exhibits itself WRT small writes and audio stutter.

(This will need to be tested empirically, and will likely not be
merged for some time. Submitted here for evaluation and recording.)

Change-Id: Ic8ce8befac7b0cbb40237961bd5d9d4cf70c13ac